day limit

+++ src/edu/sjsu/cmpe275/project/dao/CheckoutDao.java
@@ -11,6 +11,6 @@ public interface CheckoutDao {
 
 	public void createcheckout(Checkout checkout);
 	public List<Book> getbooks(int userid);
-	public BigInteger getbooksday(int userid);
+	public int getbooksday(int userid);
 	public List<Date> getdates(int userid);
 }
